@@ -5,10 +5,8 @@
   "repository": "shinnn/read-multiple-files",
   "author": "Shinnosuke Watanabe (https://github.com/shinnn)",
   "scripts": {
-    "pretest": "dependency-check . && dependency-check . --unused --no-dev && eslint --config @shinnn/node index.js test.js",
-    "test": "node test.js | tap-spec",
-    "coverage": "istanbul cover test.js",
-    "coveralls": "${npm_package_scripts_coverage} && istanbul-coveralls"
+    "pretest": "eslint --fix --format=codeframe index.js test.js",
+    "test": "nyc --reporter=text --reporter=html node test.js"
   },
   "license": "MIT",
   "files": [
@@ -27,17 +25,18 @@
     "asynchronously"
   ],
   "dependencies": {
-    "graceful-fs": "^4.1.2",
-    "run-parallel": "^1.1.2",
-    "strip-bom": "^2.0.0"
+    "graceful-fs": "^4.1.11",
+    "run-parallel": "^1.1.6",
+    "strip-bom": "^3.0.0",
+    "strip-bom-buf": "^1.0.0"
   },
   "devDependencies": {
-    "@shinnn/eslint-config-node": "^1.0.1",
-    "dependency-check": "^2.5.1",
-    "eslint": "^1.6.0",
-    "istanbul": "^0.3.22",
-    "istanbul-coveralls": "^1.0.3",
-    "tap-spec": "^4.1.0",
-    "tape": "^4.2.1"
+    "@shinnn/eslint-config-node": "^4.0.2",
+    "eslint": "^4.6.1",
+    "nyc": "^11.2.1",
+    "tape": "^4.8.0"
+  },
+  "eslintConfig": {
+    "extends": "@shinnn/node"
   }
 }
