@@ -4,9 +4,10 @@
 */
 'use strict';
 
-const fs = require('graceful-fs');
+const readFile = require('graceful-fs').readFile;
 const runParalell = require('run-parallel');
 const stripBom = require('strip-bom');
+const stripBomBuf = require('strip-bom-buf');
 
 module.exports = function readMultipleFiles(filePaths, options, cb) {
   if (cb === undefined) {
@@ -15,25 +16,31 @@ module.exports = function readMultipleFiles(filePaths, options, cb) {
   }
 
   if (typeof cb !== 'function') {
-    throw new TypeError(
-      String(cb) +
-      ' is not a function. Last argument to read-multiple-files must be a callback function.'
-    );
+    throw new TypeError(String(cb) +
+      ' is not a function. Last argument to read-multiple-files must be a callback function.');
   }
 
   if (!Array.isArray(filePaths)) {
-    throw new TypeError(
-      String(filePaths) +
-      ' is not an array. First Argument to read-multiple-files must be an array of file paths.'
-    );
+    throw new TypeError(String(filePaths) +
+      ' is not an array. First Argument to read-multiple-files must be an array of file paths.');
   }
 
-  runParalell(filePaths.map(filePath => fs.readFile.bind(fs, filePath, options)), (err, result) => {
+  runParalell(filePaths.map(filePath => done => readFile(filePath, options, done)), (err, results) => {
     if (err) {
       cb(err);
       return;
     }
 
-    cb(null, stripBom(result));
+    if (results.length === 0) {
+      cb(null, results);
+      return;
+    }
+
+    if (Buffer.isBuffer(results[0])) {
+      cb(null, results.map(stripBomBuf));
+      return;
+    }
+
+    cb(null, results.map(stripBom));
   });
 };
